ben/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
dj-database-url==2.1.0
ruff==0.12.8
